requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
pytest
httpx
jinja2
